+++ dataContainers.py
@@ -1,5 +1,5 @@
 class basePromoData:
-    def __init__(self, titleEntry, descEntry, wrestlersEntry, team1Var, team2Var, refVar):
+    def __init__(self, titleEntry, descEntry, wrestlersEntry, team1Var, team2Var, refVar, useNamesVar, surpirseEntrantsEntry, nextPromoEntry):
         self.titleEntry = titleEntry
         self.descEntry = descEntry
         self.wrestlersEntry = wrestlersEntry
@@ -7,6 +7,9 @@ class basePromoData:
         self.team2Var = team2Var
         self.refVar = refVar
         self.fileDirectory = '.'
+        self.useNamesVar = useNamesVar
+        self.surpriseEntrantsEntry = surpirseEntrantsEntry
+        self.nextPromoEntry = nextPromoEntry
     
 class singlePageData:
     def __init__(self, line1Entry, line2Entry, speakerEntry, receiverEntry, tauntEntry, demeanorEntry, featureData):
@@ -17,6 +20,7 @@ class singlePageData:
         self.tauntEntry = tauntEntry
         self.demeanorEntry = demeanorEntry
         self.featureData = []
+        self.featurePos = 0;
         
 class singleFeatureData:
     def __init__(self, featureEntry):