+++ features.py
@@ -15,17 +15,18 @@ def updateID(idFrame, basePromoData):
         for widgets in idFrame.winfo_children():
             widgets.destroy()
         for i in range(wrestlers):
-            tk.Label(idFrame,text=f"Wrestler {i+1} id: {i+1}").grid()
+            tk.Label(idFrame,text=f"Wrestler {i+1} speaker id: {i+1}").grid()
         if team1 == 1:
-            tk.Label(idFrame,text=f"Wrestler 1 team member id: {wrestlers + team1}").grid()
+            tk.Label(idFrame,text=f"Wrestler 1 team member speaker id: {wrestlers + team1}").grid()
         if team2 == 1:
-            tk.Label(idFrame,text=f"Wrestler 2 team member id: {wrestlers + team1 + team2}").grid()
+            tk.Label(idFrame,text=f"Wrestler 2 team member speaker id: {wrestlers + team1 + team2}").grid()
         if ref == 1:
-            tk.Label(idFrame,text=f"Ref id: {wrestlers + team1 + team2 + ref}").grid()
+            tk.Label(idFrame,text=f"Ref speaker id: {wrestlers + team1 + team2 + ref}").grid()
         
     except:
-        traceback.print_exc()
-        showerror(title='Wrestling Empire Custom Promo Generator', message="An error has occured, check if everything is correct and try again!")
+        with open("log.txt", "w") as log:
+            traceback.print_exc(file=log)
+            showerror(title='Wrestling Empire Custom Promo Generator', message="An error has occured, check if everything is correct and try again!")
 
 def newPage(frame, pages):
     newPageframe = tk.Frame(frame)
@@ -63,7 +64,7 @@ def newPage(frame, pages):
     pages.append(pageInfo)
     
     
-    addNewFeatureButton = tk.Button(newPageframe, text="Add new feature", command=lambda:[newFeature(newPageframe, pageInfo.featureData)])
+    addNewFeatureButton = tk.Button(newPageframe, text="Add new feature", command=lambda:[newFeature(newPageframe, pageInfo)])
     addNewFeatureButton.grid(row=13)
     
     FrameFooter = tk.Frame(newPageframe)
@@ -77,19 +78,21 @@ def newPage(frame, pages):
     
     return newPageframe, featureData
 
-def newFeature(frame, features):
+def newFeature(frame, pageInfo):
     newFeatureframe = tk.Frame(frame)
     featureLabel = tk.Label(newFeatureframe, text="Enter the promo feature")
     featureLabel.grid()
     featureEntry = tk.Entry(newFeatureframe, width=50)
     featureEntry.grid()
     featureInfo = dataContainers.singleFeatureData(featureEntry)
-    features.append(featureInfo)
-    deleteFeatureButton = tk.Button(newFeatureframe, text="Delete this feature", command=lambda:[lambda: features.remove(featureInfo),newFeatureframe.destroy()])
+    pageInfo.featureData.append(featureInfo)
+    deleteFeatureButton = tk.Button(newFeatureframe, text="Delete this feature", command=lambda:[pageInfo.featureData.remove(featureInfo), newFeatureframe.destroy()])
     deleteFeatureButton.grid()
-    newFeatureframe.grid(row=len(features)+13)
+    newFeatureframe.grid(row=pageInfo.featurePos+14)
+    pageInfo.featurePos += 1
 def formatPromo(basePromoData, pages):
     try:
+        promoFileText = ""
         title = basePromoData.titleEntry.get()
         if title == "":
             raise Exception("No title")
@@ -103,36 +106,31 @@ def formatPromo(basePromoData, pages):
         team2 = basePromoData.team2Var.get()
         ref = basePromoData.refVar.get()
         total = wrestlers + team1 + team2 + ref
+        names = basePromoData.useNamesVar.get()
+        surprises = basePromoData.surpriseEntrantsEntry.get()
         if len(pages) == 0:
             raise Exception("No pages")
-        for x in pages:
-            if (x.speakerEntry.get() == "" or x.receiverEntry.get() == ""):
-                raise Exception("No speaker/target")
-        filetypes = [('.promo files', '*.promo'),
-                    ('All files', '*.*')]
-        filename = fd.asksaveasfilename(
-            title='Save a file',
-            initialdir=basePromoData.fileDirectory,
-            filetypes=filetypes,
-            initialfile=title + ".promo")
-        basePromoData.fileDirectory = os.path.split(filename)[0]
-        f = open(filename, "w")
-        f.write("title: " + title + "\n")
-        f.write("description: " + description + "\n")
-        f.write("characters: ")
+
+        promoFileText += "title: " + title + "\n"
+        promoFileText += "description: " + description + "\n"
+        promoFileText += "characters: "
         for i in range(wrestlers):
             if (i != 0):
-                f.write(",")
-            f.write(str(i+1))
+                promoFileText += ","
+            promoFileText += str(i+1)
         if (team1 == 1):
-            f.write(",11")
+            promoFileText += ",11"
         if (team2 == 1):
-            f.write(",22")
+            promoFileText += ",22"
         if(wrestlers != 0 and ref == 1):
-            f.write(",")
+            promoFileText += ","
         if (ref == 1):
-            f.write("-1")
-        f.write("\n")
+            promoFileText += "-1"
+        promoFileText += "\n"
+        if (names == 1):
+            promoFileText += "use_names: True\n"
+        if (len(surprises)!=0):
+            promoFileText += "surprise_entrants: " + surprises + "\n"
         for x in pages:
             strline = '"' + x.line1Entry.get() + '","' + x.line2Entry.get() + '",' + x.speakerEntry.get() + ',' + x.receiverEntry.get() + ',' + x.tauntEntry.get() + ','
             d = 0
@@ -143,20 +141,31 @@ def formatPromo(basePromoData, pages):
             if (x.demeanorEntry.get() == "0"):
                 d = 0
             strline += str(d)
-            f.write(strline)
+            promoFileText += strline
             if x.featureData:
-                f.write(',')
+                promoFileText += ','
                 feature = ""
                 for y in x.featureData:
                     feature += y.featureEntry.get() + ';'
                 feature = feature.strip(';')    
-                f.write(feature)
-            f.write("\n")
+                promoFileText += feature
+            promoFileText += "\n"
+        filetypes = [('.promo files', '*.promo'),
+                    ('All files', '*.*')]
+        filename = fd.asksaveasfilename(
+            title='Save a file',
+            initialdir=basePromoData.fileDirectory,
+            filetypes=filetypes,
+            initialfile=title + ".promo")
+        basePromoData.fileDirectory = os.path.split(filename)[0]
+        f = open(filename, "w")
+        f.write(promoFileText)
         f.close()
         showinfo(title='Wrestling Empire Custom Promo Generator', message=".promo file was successfully generated! You can now close this tool")
-    except:
-        traceback.print_exc()
-        showerror(title='Wrestling Empire Custom Promo Generator', message="An error has occured, check if everything is correct and try again!")
+    except Exception as e:
+        with open("log.txt", "w") as log:
+            traceback.print_exc(file=log)
+            showerror(title='Wrestling Empire Custom Promo Generator', message="An error has occured, check if everything is correct and try again!")
 
 def cleanUp(basePromoData, pages, idFrame, canvas):
     canvas.yview_moveto(0)
@@ -166,6 +175,8 @@ def cleanUp(basePromoData, pages, idFrame, canvas):
     basePromoData.team1Var.set(0)
     basePromoData.team2Var.set(0)
     basePromoData.refVar.set(0)
+    basePromoData.useNamesVar.set(0)
+    basePromoData.surpriseEntrantsEntry.delete(0, tk.END)
     for widgets in idFrame.winfo_children():
         widgets.destroy()
     tk.Label(idFrame,text="No wrestlers").grid()
@@ -188,10 +199,14 @@ def loadPromo(basePromoData, pages, idFrame, frame, canvas):
         f = open(filename, "r")
         lines = f.readlines()
         f.close()
+        promoFileLine = 0;
         basePromoData.fileDirectory = os.path.split(filename)[0]
-        basePromoData.titleEntry.insert(-1, lines[0][7:].strip())
-        basePromoData.descEntry.insert(-1, lines[1][13:].strip())
-        wrestlersLine = lines[2][12:].strip().split(',')
+        basePromoData.titleEntry.insert(-1, lines[promoFileLine][7:].strip())
+        promoFileLine+=1
+        basePromoData.descEntry.insert(-1, lines[promoFileLine][13:].strip())
+        promoFileLine+=1
+        wrestlersLine = lines[promoFileLine][12:].strip().split(',')
+        promoFileLine+=1
         wrestlers = 0
         team1 = 0
         team2 = 0
@@ -216,8 +231,14 @@ def loadPromo(basePromoData, pages, idFrame, frame, canvas):
         basePromoData.team2Var.set(team2)
         basePromoData.refVar.set(ref)
         updateID(idFrame, basePromoData)
-
-        for i, x in enumerate(lines[3:]):
+        if(lines[promoFileLine][0:10] == "use_names:"):
+            if(lines[promoFileLine][11:].strip().lower() == "true"):
+                basePromoData.useNamesVar.set(1)
+                promoFileLine+=1
+        if(lines[promoFileLine][0:18] == "surprise_entrants:"):
+            basePromoData.surpriseEntrantsEntry.insert(-1, lines[promoFileLine][19:].strip())
+            promoFileLine+=1
+        for i, x in enumerate(lines[promoFileLine:]):
             newPageFrame, featureData = newPage(frame, pages)
             pageLine = re.split(r',(?=(?:[^"]|"[^"\\]*(?:\\\\.[^"\\]*)*")*$)', x) #regex source: chatgpt
             line1 = pageLine[0][1:-1]
@@ -244,9 +265,10 @@ def loadPromo(basePromoData, pages, idFrame, frame, canvas):
             if(len(pageLine) > 6):
                 featureLine = pageLine[6].split(';')
                 for j, y in enumerate(featureLine):
-                    newFeature(newPageFrame, pages[i].featureData)
+                    newFeature(newPageFrame, pages[i])
                     pages[i].featureData[j].featureEntry.insert(-1, featureLine[j].strip())
                     
     except:
-        traceback.print_exc()
-        showerror(title='Wrestling Empire Custom Promo Generator', message="Failed to load the file.")
+        with open("log.txt", "w") as log:
+            traceback.print_exc(file=log)
+            showerror(title='Wrestling Empire Custom Promo Generator', message="Failed to load the file.")