+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## 1.3.0
+- WECCL x.x.x: now supports surprise entrants and using names instead of ids.
+- Fixed the issue that happens when you delete a promo feature
+- The tool now creates log.txt file when it fails to save the promo file.
+- The tool should now avoid making incorrect .promo files when it fails to save them properly.
 ## 1.2.0
 - WECCL 1.2.5: Now supports new advanced promo features; now also accepts taunt names instead of numbers; using `$promotion#`, `$$belt(#1)_(#2)`, `$$champ(#1)_(#2)` in promo lines.
 - Saving to a different location